Move cnodes module to higher level (out of uflacs).
And cleanup some imports.
