+++ setup.py
@@ -125,10 +125,10 @@ def run_install():
             "ffc.backends",
             "ffc.backends.ffc",
             "ffc.backends.ufc",
+            "ffc.language",
             "ffc.tsfc",
             "ffc.uflacs",
             "ffc.uflacs.analysis",
-            "ffc.uflacs.language",
         ],
         package_dir={"ffc": "ffc"},
         package_data={"ffc": [os.path.join('backends', 'ufc', '*.h')]},
